requirements.txt
flask==1.1.1
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.27
PyJWT
psycopg2-binary
flask-migrate==2.5.2